Fragments: add HomeFragment and HomeViewModel
the architecture I am trying to achieve here is.

MainActivity holds the bottomsheet with all the menu options and each page is a different fragment, bar a few pages like messages. This makes the ui incredibly flexible as we can just add a new fragment should we need a new screen.